Move swarm session chips into the selected journey host cell.
Share run session state via a provider so the matrix renders under the focused target in JourneyBlock while the right panel keeps the live stream.
